bisection: Add an initial script for future bisection exercises.
The idea of the script is to count how many lines of comments are
there in the repo. So over time, when the repo grows, one comment
will make it over the set limit, to trigger the alert. This is
when you can start to use `git bisect` to find out the culprit.
In this way, you won't actually know which commit triggers the
exception by just looking at the commit contents (hopefully!)

Also, just in case you want to start earlier, or the number of
commits never go above the default value (1000), the script also
takes an argument to change the limit, so we can start the
exercise anytime.

As long as python is installed on your machine, the script should
be able to execute. Will seek other options if that's not working